verwaltung changes round 2

+++ tfmarkt/Produktklassen/Produkt.cs
@@ -9,7 +9,7 @@ namespace tfmarkt.Produktklassen
     //Abstrakte Klasse für alle Produkte
     public abstract class Produkt 
     {
-        public double preis { get; set; }
+        public decimal preis { get; set; }
         public string name { get; set; }
         public int artikelnummer { get; set; }
         public string beschreibung { get; set; }
